+++ ntia_conformance_checker/fsct_checker.py
@@ -186,7 +186,7 @@ class FSCT3Checker(BaseChecker):
                     "The following errors were found:\n"
                 )
                 for message in self.validation_messages:
-                    print(message)
+                    print(message.validation_message)
 
     def output_json(self) -> Dict[str, Any]:
         """Create a dict of results for outputting to JSON."""
@@ -284,7 +284,7 @@ class FSCT3Checker(BaseChecker):
                     "The following errors were found:</p>\n"
                 )
                 for message in self.validation_messages:
-                    result += f"<p>{message}</p>\n"
+                    result += f"<p>{message.validation_message}</p>\n"
         else:
             result = f"""
             <h2>FSCTv3-Minimum Expected Conformance Results</h2>